[DPM-133] Move service file to separated file
